client/: move RpcClient from rpc_request.rs to rpc_client.rs
