+++ client/src/lib.rs
@@ -1,5 +1,6 @@
 mod generic_rpc_client_request;
 pub mod mock_rpc_client_request;
+pub mod rpc_client;
 pub mod rpc_client_request;
 pub mod rpc_request;
 pub mod rpc_signature_status;
