+++ client/src/thin_client.rs
@@ -3,7 +3,7 @@
 //! messages to the network directly. The binary encoding of its messages are
 //! unstable and may change in future releases.
 
-use crate::rpc_request::RpcClient;
+use crate::rpc_client::RpcClient;
 use bincode::serialize_into;
 use log::*;
 use solana_metrics;
