+++ core/src/replicator.rs
@@ -12,7 +12,8 @@ use crate::storage_stage::{get_segment_from_entry, ENTRIES_PER_SEGMENT};
 use crate::window_service::WindowService;
 use rand::thread_rng;
 use rand::Rng;
-use solana_client::rpc_request::{RpcClient, RpcRequest};
+use solana_client::rpc_client::RpcClient;
+use solana_client::rpc_request::RpcRequest;
 use solana_client::thin_client::{create_client, retry_get_balance, ThinClient};
 use solana_drone::drone::{request_airdrop_transaction, DRONE_PORT};
 use solana_sdk::hash::{Hash, Hasher};
