+++ wallet/tests/deploy.rs
@@ -1,6 +1,7 @@
 use serde_json::{json, Value};
 use solana::fullnode::new_fullnode_for_tests;
-use solana_client::rpc_request::{RpcClient, RpcRequest};
+use solana_client::rpc_client::RpcClient;
+use solana_client::rpc_request::RpcRequest;
 use solana_drone::drone::run_local_drone;
 use solana_sdk::bpf_loader;
 use solana_wallet::wallet::{process_command, WalletCommand, WalletConfig};
