+++ wallet/tests/pay.rs
@@ -1,6 +1,6 @@
 use chrono::prelude::*;
 use serde_json::Value;
-use solana_client::rpc_request::RpcClient;
+use solana_client::rpc_client::RpcClient;
 use solana_drone::drone::run_local_drone;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::{Keypair, KeypairUtil};
